dev/requirements-dev.txt
datasets
pillow
pytesseract
tqdm
